requirements.txt
boto3
pandas
tensorflow
Pillow
numpy
opencv-contrib-python-headless
torch
torchvision
ultralytics
supervision
openai
streamlit-drawable-canvas
